新增SkillManager 可以释放技能，并在UI上显示

+++ Client/Assets/Scripts/Manager/DataManager.cs
@@ -138,7 +138,7 @@ public class DataManager : Singleton<DataManager>
         }
         return data;
     }
-    private int GetPlayerHp()
+    private float GetPlayerHp()
     {
         return UserManager.Instance.HP;
     }
